chore(refactor): rename local-data to config for clarity (refs #47)
