+++ src/overwrite/overwrite-store.ts
@@ -2,7 +2,7 @@ import { existsSync, readFileSync } from 'fs';
 import { join } from 'path';
 
 import type { CollectionOverwrite } from '../types';
-import { getDataDir } from '../local-data/data-dir';
+import { getDataDir } from '../config/data-dir';
 import { parseOverwrite } from './overwrite';
 
 /*
